+++ poc/classes/AuxSTConstructor.py
@@ -1,11 +1,11 @@
 from poc.classes.AuxInbuiltTypes import InbuiltClassInstance, InbuiltUndefined
-from poc.classes.AuxSTBlock import AuxSTBlock
+from poc.classes.AuxSTBuildingBlock import AuxSTBuildingBlock
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 from anytree import search
 from poc.fplerror import FplErrorManager
 
 
-class AuxSTConstructor(AuxSTBlock):
+class AuxSTConstructor(AuxSTBuildingBlock):
 
     def __init__(self, i):
         super().__init__(AuxSymbolTable.classConstructor, i)