+++ poc/classes/AuxSTTheoremLikeStatementOrConjecture.py
@@ -1,6 +1,6 @@
 from poc.classes.AuxEvaluation import EvaluateParams
 from poc.classes.AuxInbuiltTypes import InbuiltPredicate, InbuiltUndefined, EvaluatedPredicate
-from poc.classes.AuxSTBlock import AuxSTBlock
+from poc.classes.AuxSTBuildingBlock import AuxSTBuildingBlock
 from poc.classes.AuxST import AuxSTOutline
 from poc.classes.AuxSTPredicate import AuxSTPredicate
 from poc.classes.AuxSTSignature import AuxSTSignature
@@ -8,7 +8,7 @@ from poc.classes.AuxSTVarSpecList import AuxSTVarSpecList
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 
 
-class AuxSTTheoremLikeStatementOrConjecture(AuxSTBlock):
+class AuxSTTheoremLikeStatementOrConjecture(AuxSTBuildingBlock):
 
     def __init__(self, block_type, i, zfrom, zto):
         super().__init__(block_type, i)